+++ javascript/user-interface/hover-markers/CelestialHoverMarker.jsx
@@ -107,12 +107,14 @@ class CelestialObjectMarker extends Component {
 
     HandleMouseEnter() {
         this.setState({ isHovering: true });
+        this.celestialObjectDelegate.OnHoverEnter();
 
         GameManager.gameObserver.Dispatch("OnHoverMarkerEnter");
     }
 
     HandleMouseExit() {
         this.setState({ isHovering: false });
+        this.celestialObjectDelegate.OnHoverExit();
 
         GameManager.gameObserver.Dispatch("OnHoverMarkerExit");
     }
@@ -140,7 +142,7 @@ class CelestialObjectMarker extends Component {
                         left: `${this.state.screenPosition.x - elementHalfWidth}px`,
                         opacity: shouldRender ? "1" : "0",
                         pointerEvents: shouldRender ? "all" : "none",
-                        backgroundColor: this.state.color
+                        backgroundColor: this.state.isHovering ? "transparent" : this.state.color
                     }}
                     onClick={this.HandleClick.bind(this)}
                     onMouseEnter={this.HandleMouseEnter.bind(this)}